split queries into several files
