+++ dev/archiveTest.mts
@@ -1,6 +1,6 @@
 import archiveImport from '../store/archive.js';
+import { readArchivedMatch } from '../store/archiveHelpers.js';
 import {
-  readArchivedMatch,
   getMatchDataFromCassandra,
   getPlayerMatchData,
 } from '../store/queries.js';
